backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.11.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
groq==0.9.0
httpx==0.27.0
alembic==1.13.1
slowapi==0.1.9